Fixed JQuery clone problem.
If you've ever used .clone() in jQuery, you have encountered an issue where buttons become unresponsive after cloning. This occurs because the cloned element is already rendered, and that's what we need to address. To resolve this, we've implemented a check to determine if an element has already been rendered. If it has, we remove all elements added in the previous call to the .render() function and then proceed to render it once more.